+++ packages/kit/src/vite/index.js
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import colors from 'kleur';
 import { svelte } from '@sveltejs/vite-plugin-svelte';
-import { searchForWorkspaceRoot } from 'vite';
+import * as vite from 'vite';
 import { mkdirp, posixify, rimraf } from '../utils/filesystem.js';
 import * as sync from '../core/sync/sync.js';
 import { build_server } from './build/build_server.js';
@@ -12,7 +12,7 @@ import { load_config } from '../core/config/index.js';
 import { dev } from './dev/index.js';
 import { generate_manifest } from '../core/generate_manifest/index.js';
 import { get_runtime_directory, logger } from '../core/utils.js';
-import { find_deps, get_default_config } from './build/utils.js';
+import { find_deps, get_default_config as get_default_build_config } from './build/utils.js';
 import { preview } from './preview/index.js';
 import { get_aliases, resolve_entry } from './utils.js';
 
@@ -111,7 +111,7 @@ function kit() {
 
 	let completed_build = false;
 
-	function vite_client_config() {
+	function vite_client_build_config() {
 		/** @type {Record<string, string>} */
 		const input = {
 			// Put unchanging assets in immutable directory. We don't set that in the
@@ -132,7 +132,7 @@ function kit() {
 			input[name] = resolved;
 		});
 
-		return get_default_config({
+		return get_default_build_config({
 			config: svelte_config,
 			input,
 			ssr: false,
@@ -180,17 +180,13 @@ function kit() {
 			paths = {
 				build_dir: `${svelte_config.kit.outDir}/build`,
 				output_dir: `${svelte_config.kit.outDir}/output`,
-				client_out_dir: `${svelte_config.kit.outDir}/output/client/${svelte_config.kit.appDir}`
+				client_out_dir: `${svelte_config.kit.outDir}/output/client/`
 			};
 
 			if (is_build) {
-				process.env.VITE_SVELTEKIT_APP_VERSION = svelte_config.kit.version.name;
-				process.env.VITE_SVELTEKIT_APP_VERSION_FILE = `${svelte_config.kit.appDir}/version.json`;
-				process.env.VITE_SVELTEKIT_APP_VERSION_POLL_INTERVAL = `${svelte_config.kit.version.pollInterval}`;
-
 				manifest_data = sync.all(svelte_config).manifest_data;
 
-				const new_config = vite_client_config();
+				const new_config = vite_client_build_config();
 
 				warn_overridden_config(config, new_config);
 
@@ -209,6 +205,10 @@ function kit() {
 						input: `${get_runtime_directory(svelte_config.kit)}/client/start.js`
 					}
 				},
+				define: {
+					__SVELTEKIT_DEV__: 'true',
+					__SVELTEKIT_APP_VERSION_POLL_INTERVAL__: '0'
+				},
 				resolve: {
 					alias: get_aliases(svelte_config.kit)
 				},
@@ -222,7 +222,7 @@ function kit() {
 								svelte_config.kit.outDir,
 								path.resolve(cwd, 'src'),
 								path.resolve(cwd, 'node_modules'),
-								path.resolve(searchForWorkspaceRoot(cwd), 'node_modules')
+								path.resolve(vite.searchForWorkspaceRoot(cwd), 'node_modules')
 							])
 						]
 					},
@@ -272,8 +272,8 @@ function kit() {
 			});
 
 			fs.writeFileSync(
-				`${paths.client_out_dir}/version.json`,
-				JSON.stringify({ version: process.env.VITE_SVELTEKIT_APP_VERSION })
+				`${paths.client_out_dir}/${svelte_config.kit.appDir}/version.json`,
+				JSON.stringify({ version: svelte_config.kit.version.name })
 			);
 
 			const { assets, chunks } = collect_output(bundle);
@@ -317,8 +317,8 @@ function kit() {
 
 			const files = new Set([
 				...static_files,
-				...chunks.map((chunk) => `${svelte_config.kit.appDir}/${chunk.fileName}`),
-				...assets.map((chunk) => `${svelte_config.kit.appDir}/${chunk.fileName}`)
+				...chunks.map((chunk) => chunk.fileName),
+				...assets.map((chunk) => chunk.fileName)
 			]);
 
 			// TODO is this right?
@@ -402,18 +402,14 @@ function kit() {
 }
 
 function check_vite_version() {
-	let vite_major = 3;
+	// TODO parse from kit peer deps and maybe do a full semver compare if we ever require feature releases a min
+	const min_required_vite_major = 3;
+	const vite_version = vite.version ?? '2.x'; // vite started exporting it's version in 3.0
+	const current_vite_major = parseInt(vite_version.split('.')[0], 10);
 
-	try {
-		const pkg = JSON.parse(fs.readFileSync('package.json', 'utf-8'));
-		vite_major = +pkg.devDependencies['vite'].replace(/^[~^]/, '')[0];
-	} catch {
-		// do nothing
-	}
-
-	if (vite_major < 3) {
+	if (current_vite_major < min_required_vite_major) {
 		throw new Error(
-			`Vite version ${vite_major} is no longer supported. Please upgrade to version 3`
+			`Vite version ${current_vite_major} is no longer supported. Please upgrade to version ${min_required_vite_major}`
 		);
 	}
 }