+++ packages/kit/test/apps/basics/test/server.test.js
@@ -1,5 +1,6 @@
 import { expect } from '@playwright/test';
 import { start_server, test } from '../../../utils.js';
+import { createHash, randomBytes } from 'node:crypto';
 
 /** @typedef {import('@playwright/test').Response} Response */
 
@@ -24,6 +25,13 @@ test.describe('Caching', () => {
 		request
 	}) => {
 		const response = await request.get('/caching/private/uses-session-in-init');
+		expect(response.headers()['cache-control']).toBe('private, max-age=30');
+	});
+
+	test('sets cache-control: private even if page doesnt use session but one exists and cache.private is unset', async ({
+		request
+	}) => {
+		const response = await request.get('/caching/private/has-session');
 		expect(response.headers()['cache-control']).toBe('private, max-age=30');
 	});
 
@@ -62,6 +70,13 @@ test.describe('Caching', () => {
 		request
 	}) => {
 		const response = await request.get('/caching/private/uses-session-in-init?private=false');
+		expect(response.headers()['cache-control']).toBe('public, max-age=30');
+	});
+
+	test('sets cache-control: public if page has session and cache.private is false', async ({
+		request
+	}) => {
+		const response = await request.get('/caching/private/has-session?private=false');
 		expect(response.headers()['cache-control']).toBe('public, max-age=30');
 	});
 
@@ -271,9 +286,20 @@ test.describe('Endpoints', () => {
 	});
 
 	test('body can be a binary ReadableStream', async ({ request }) => {
+		const interruptedResponse = request.get('/endpoint-output/stream-throw-error');
+		await expect(interruptedResponse).rejects.toThrow('socket hang up');
+
 		const response = await request.get('/endpoint-output/stream');
 		const body = await response.body();
-		expect(Array.from(body)).toEqual([1, 2, 3]);
+		const digest = createHash('sha256').update(body).digest('base64url');
+		expect(response.headers()['digest']).toEqual(`sha-256=${digest}`);
+	});
+
+	test('request body can be read slow', async ({ request }) => {
+		const data = randomBytes(1024 * 256);
+		const digest = createHash('sha256').update(data).digest('base64url');
+		const response = await request.put('/endpoint-input/sha256', { data });
+		expect(await response.text()).toEqual(digest);
 	});
 });