+++ packages/adapter-vercel/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-vercel
 
+## 1.0.0-next.65
+
+### Patch Changes
+
+- [breaking] remove writeStatic to align with Vite ([#5618](https://github.com/sveltejs/kit/pull/5618))
+
+## 1.0.0-next.64
+
+### Patch Changes
+
+- Don't remove .vercel/output directory ([#5555](https://github.com/sveltejs/kit/pull/5555))
+
 ## 1.0.0-next.63
 
 ### Patch Changes